Finished Wisdom branch
Still working on courage and power branches, got to second branch. Plan the first letter of each description in chronological order to spell out the following; Im a hostage and they force me to work. "im a hostage" already in wisdom branch, hence the other branches would be "and they for" "ce me to work".